smartapi_ma_crossover_alert/requirements.txt
smartapi-python
requests
python-dotenv
pandas
numpy
pyotp
pytz
schedule
